+++ businessLogic/folder/getAcessTokenFromCookie.ts
