+++ businessLogic/folder/getFolderIdFromQuery.ts
